Lab03/requirements.txt
PyQt5
request
Flask>=2.3.2
rsa>=4.9
ecdsa